Split client side rendering into controller and view
