+++ zapzap/views/client_side_rendering/__init__.py
@@ -1,0 +1,1 @@
+from .client_side_rendering_view import ClientSideRenderingView
